actualización código atenas y fallo dato int32 que era double

+++ PROYECTO_LENGUAJE_MARCAS/Models/apiAqi.cs
@@ -4,13 +4,13 @@ public class ApiAqi
 {
      public class Datum
     {
-        public int aqi { get; set; }
+        public double aqi { get; set; }
         public double o3 { get; set; }
         public double so2 { get; set; }
         public double no2 { get; set; }
         public double co { get; set; }
-        public int pm10 { get; set; }
-        public int pm25 { get; set; }
+        public double pm10 { get; set; }
+        public double pm25 { get; set; }
         public int pollen_level_tree { get; set; }
         public int pollen_level_grass { get; set; }
         public int pollen_level_weed { get; set; }
